[#90886] - Added implementation for adding be default blocked road sections (not traversable sections based on nwb database in comparison with traversal through graphhopper

+++ accessibility/src/main/java/nu/ndw/nls/accessibilitymap/accessibility/services/AccessibilityMapService.java
@@ -163,7 +163,7 @@ public class AccessibilityMapService {
     }
 
     private RoadSection initializeRoadSection(AccessibilityRoadSection accessibilityRoadSection) {
-        return new RoadSection(accessibilityRoadSection.getRoadSectionId(),
+        return new RoadSection((int) accessibilityRoadSection.getRoadSectionId(),
                 accessibilityRoadSection.getGeometry(),
                 initializeNwbAccessibleRoads(accessibilityRoadSection.isForwardAccessible()),
                 initializeNwbAccessibleRoads(accessibilityRoadSection.isBackwardAccessible()));
